optimization of rdtsc()
On x86-64, the most significant bits of the RDX:RAX registers are
guaranteed to be zero. This means that the rdtsc instruction will return
two 64-bit numbers. In practice, this removes the extra instruction to
clear the most significant bits of the return value.

+++ runtime/Cycles.h
@@ -42,7 +42,7 @@ class Cycles {
         if (mockTscValue)
             return mockTscValue;
 #endif
-        uint32_t lo, hi;
+        size_t lo, hi;
         __asm__ __volatile__("rdtsc" : "=a" (lo), "=d" (hi));
 //        __asm__ __volatile__("rdtscp" : "=a" (lo), "=d" (hi) : : "%rcx");
         return (((uint64_t)hi << 32) | lo);
